refactor(backend): change lat/lon types from Float to Double

+++ backend/src/main/java/com/davisiqueira/fraud_guard/dto/transaction/TransactionRequestDTO.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record TransactionRequestDTO(
         String cpf,
         BigDecimal value,
-        Float lat,
-        Float lon
+        Double lat,
+        Double lon
 ) {
 }
